+++ NEWS.md
@@ -8,6 +8,7 @@
 * Fixed the error message for unexpected Multiplicities in the input and output type checking during `PipeOp`s training and prediction.
 * Untrained `PipeOp`s that take `NULL` as input during training now automatically perform training during prediction.
 * `PipeOpImputeConstant`, `PipeOpImputeMode`, `PipeOpImputeOOR`, and `PipeOpImputeLearner` can now handle `factor` or `ordered` features with zero levels.
+* Removed compatibility for old `paradox` versions pre-1.0.0.
 
 # mlr3pipelines 0.8.0